Fixed remaining code smells
The main challenge was addressing a few constructors we had that accepted 10 or so arguments. To address that in each case, I moved all the args into a separate class, and where appropriate, I added "with" methods to allow an instance of the class to be easily constructed and populated with only the desired args.