Now supporting SSL
As noted in the comments, I think we can make this better once Java Client 6.4.0 is released with configurable support for 2-way SSL. 

No automated tests yet. It's simple to add a new app server to the test-app that requires SSL, but because this is using the JVM's default trust manager, the call will fail since the JVM won't trust the certificate template used by MarkLogic.