Fixing Mule compiler warnings about exported packages
I noticed this warning when running "mvn clean install":

```
[WARNING] diagnostic: /Users/rudin/workspace/marklogic-mule-connector/src/main/java/com/marklogic/mule/extension/MarkLogicExtension.java:30: warning: The extension [MarkLogic] exports packages which doesn't have a defined visibility, 'api' or 'internal':
public class MarkLogicExtension {
       ^
  -> [com.marklogic.mule.extension]: because of these classes: 
          * com.marklogic.mule.extension.QueryType
          * com.marklogic.mule.extension.AuthenticationType
          * com.marklogic.mule.extension.DocumentAttributes
          * com.marklogic.mule.extension.ConnectionType
          * com.marklogic.mule.extension.QueryFormat
```

So I moved all of those classes to an "api" package - except for the Java Client Format class. Not sure what to do about that warning yet - we may just have to duplicate that with our own "Format" class that goes into an "api" package as well.