export-documents now supports everything read-documents does
Added `PagingProvider` classes to capture the shared and specialized logic. 

Removed the OBE `write-batch` operation, though that's still used in the `manual-test-flow` file. Will update that soon to use `write-documents`. 

I'd like to do some more refactoring later - I think Sonar is going to complain about the number of constructor arguments and the length of some methods. But wanted to get this in place first and will then refactor.