DEVEXP-648 Fixed remaining Sonar issues by using ParameterGroup
Created WriteParameters / QueryParameters to store all the params, which avoids the warnings about having a bunch of methods. Did some initial manual testing to verify it all works fine. No Sonar warnings anymore. 

Did remove the "rest" prefix for the transform/transformParams/etc inputs as I think it's not necessary - we only have one concept of a "transform". So there are a few small changes to the test flows.